+++ docs/history.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+`1.0.0`_ (20 Nov 2015)
+----------------------
+- Add ``sprockets.http.mixins.LoggingHandler``
+- Add ``sprockets.http.mixins.ErrorLogger``
+- Add ``sprockets.http.mixins.ErrorWriter``
+
 `0.4.0`_ (24 Sep 2015)
 ----------------------
 - Run callbacks from ``application.runner_callbacks['shutdown']`` when
@@ -28,5 +34,8 @@ Release History
 
 .. _0.2.0: https://github.com/sprockets/sprockets.http/compare/0.0.0...0.2.0
 .. _0.2.1: https://github.com/sprockets/sprockets.http/compare/0.2.0...0.2.1
-.. _0.3.0: https://github.com/sprockets/sprockets.http/compare/0.2.0...0.3.0
+.. _0.2.2: https://github.com/sprockets/sprockets.http/compare/0.2.1...0.2.2
+.. _0.3.0: https://github.com/sprockets/sprockets.http/compare/0.2.2...0.3.0
 .. _0.4.0: https://github.com/sprockets/sprockets.http/compare/0.3.0...0.4.0
+.. _1.0.0: https://github.com/sprockets/sprockets.http/compare/0.4.0...1.0.0
+.. _Next Release: https://github.com/sprockets/sprockets.http/compare/1.0.0...master